Handler tests put in a separate module
